backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
pymongo==4.6.1
dnspython==2.4.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Kubernetes Client
kubernetes==28.1.0

# AWS SDK
boto3==1.34.10
botocore==1.34.10

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time Utilities
python-dateutil==2.8.2

# Monitoring & Metrics
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0

# Logging
python-json-logger==2.0.7

# Data Validation
marshmallow==3.20.1
pydantic==2.5.3

# Background Tasks
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code Quality
flake8==7.0.0
black==23.12.1
pylint==3.0.3

# API Documentation
flasgger==0.9.7.1
apispec==6.3.1

# YAML Processing
PyYAML==6.0.1

# JSON Processing
orjson==3.9.10

# System Monitoring
psutil==5.9.6

# Git Integration
GitPython==3.1.40

# Terraform Python Wrapper
python-terraform==0.10.1

# OpenTelemetry (Future Tracing)
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Rate Limiting
Flask-Limiter==3.5.0

# Caching
Flask-Caching==2.1.0

# Email (for notifications)
Flask-Mail==0.9.1

# Utilities
click==8.1.7
colorama==0.4.6
tabulate==0.9.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1